:truck: Rename the repository to fhir-synthetizer
