+++ pyproject.toml
@@ -31,7 +31,7 @@ exclude_lines = [
 
 [tool.isort]
 known_local_folder = [
-    'okapy',
+    'synthetizer',
 ]
 line_length = 100
 profile = 'black'
