+++ tests/main_test.py
@@ -1,10 +1,10 @@
-from okapy import anonymization_pipeline
+from synthetizer import anonymization_pipeline
 
 
 def test_anonymization_pipeline(mocker, patients_actifs_example: dict, test_data_path):
-    # Mock the function `fetch` in okapy.main (i.e. don't enter the function, just output the
+    # Mock the function `fetch` in synthetizer.main (i.e. don't enter the function, just output the
     # `return_value`)
-    mocker.patch("okapy.main.fetch", return_value=patients_actifs_example)
+    mocker.patch("synthetizer.main.fetch", return_value=patients_actifs_example)
 
     # Test the anonymization pipeline with the mocked values
     anonymization_pipeline(
